refactor: split datanode fiter tests to a dedicated file
