+++ tests/core/data/test_data_node.py
@@ -20,83 +20,16 @@ import src.taipy.core as tp
 from src.taipy.core._orchestrator._orchestrator_factory import _OrchestratorFactory
 from src.taipy.core.config.job_config import JobConfig
 from src.taipy.core.data._data_manager import _DataManager
-from src.taipy.core.data._filter import _FilterDataNode
 from src.taipy.core.data.data_node import DataNode
 from src.taipy.core.data.data_node_id import DataNodeId
 from src.taipy.core.data.in_memory import InMemoryDataNode
-from src.taipy.core.data.operator import JoinOperator, Operator
 from src.taipy.core.exceptions.exceptions import DataNodeIsBeingEdited, NoData
 from src.taipy.core.job.job_id import JobId
 from taipy.config import Config
 from taipy.config.common.scope import Scope
 from taipy.config.exceptions.exceptions import InvalidConfigurationId
 
-
-class FakeDataNode(InMemoryDataNode):
-    read_has_been_called = 0
-    write_has_been_called = 0
-
-    def __init__(self, config_id, **kwargs):
-        scope = kwargs.pop("scope", Scope.SCENARIO)
-        super().__init__(config_id=config_id, scope=scope, **kwargs)
-
-    def _read(self, query=None):
-        self.read_has_been_called += 1
-
-    def _write(self, data):
-        self.write_has_been_called += 1
-
-    @classmethod
-    def storage_type(cls) -> str:
-        return "fake_inmemory"
-
-    write = DataNode.write  # Make sure that the writing behavior comes from DataNode
-
-
-class FakeDataframeDataNode(DataNode):
-    COLUMN_NAME_1 = "a"
-    COLUMN_NAME_2 = "b"
-
-    def __init__(self, config_id, default_data_frame, **kwargs):
-        super().__init__(config_id, **kwargs)
-        self.data = default_data_frame
-
-    def _read(self):
-        return self.data
-
-    @classmethod
-    def storage_type(cls) -> str:
-        return "fake_df_dn"
-
-
-class FakeNumpyarrayDataNode(DataNode):
-    def __init__(self, config_id, default_array, **kwargs):
-        super().__init__(config_id, **kwargs)
-        self.data = default_array
-
-    def _read(self):
-        return self.data
-
-    @classmethod
-    def storage_type(cls) -> str:
-        return "fake_np_dn"
-
-
-class FakeListDataNode(DataNode):
-    class Row:
-        def __init__(self, value):
-            self.value = value
-
-    def __init__(self, config_id, **kwargs):
-        super().__init__(config_id, **kwargs)
-        self.data = [self.Row(i) for i in range(10)]
-
-    def _read(self):
-        return self.data
-
-    @classmethod
-    def storage_type(cls) -> str:
-        return "fake_list_dn"
+from .utils import FakeDataNode
 
 
 def funct_a_b(input: str):
@@ -451,197 +384,6 @@ class TestDataNode:
         assert ("task_a_b", tp.Status.SKIPPED) in jobs_and_status
         assert ("task_b_c", tp.Status.COMPLETED) in jobs_and_status
         assert ("task_b_d", tp.Status.COMPLETED) in jobs_and_status
-
-    def test_pandas_filter(self, default_data_frame):
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
-        COLUMN_NAME_1 = "a"
-        COLUMN_NAME_2 = "b"
-        assert isinstance(df_dn[COLUMN_NAME_1], _FilterDataNode)
-        assert isinstance(df_dn[[COLUMN_NAME_1, COLUMN_NAME_2]], _FilterDataNode)
-
-    def test_filter_pandas_exposed_type(self, default_data_frame):
-        dn = FakeDataNode("fake_dn")
-        dn.write("Any data")
-
-        with pytest.raises(NotImplementedError):
-            dn.filter((("any", 0, Operator.EQUAL)), JoinOperator.OR)
-        with pytest.raises(NotImplementedError):
-            dn.filter((("any", 0, Operator.EQUAL)), JoinOperator.OR)
-        with pytest.raises(NotImplementedError):
-            dn.filter((("any", 0, Operator.NOT_EQUAL)), JoinOperator.OR)
-        with pytest.raises(NotImplementedError):
-            dn.filter((("any", 0, Operator.LESS_THAN)), JoinOperator.AND)
-        with pytest.raises(NotImplementedError):
-            dn.filter((("any", 0, Operator.LESS_OR_EQUAL)), JoinOperator.AND)
-        with pytest.raises(NotImplementedError):
-            dn.filter((("any", 0, Operator.GREATER_THAN)))
-        with pytest.raises(NotImplementedError):
-            dn.filter(("any", 0, Operator.GREATER_OR_EQUAL))
-
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
-
-        COLUMN_NAME_1 = "a"
-        COLUMN_NAME_2 = "b"
-        assert len(df_dn.filter((COLUMN_NAME_1, 1, Operator.EQUAL))) == len(
-            default_data_frame[default_data_frame[COLUMN_NAME_1] == 1]
-        )
-        assert len(df_dn.filter((COLUMN_NAME_1, 1, Operator.NOT_EQUAL))) == len(
-            default_data_frame[default_data_frame[COLUMN_NAME_1] != 1]
-        )
-        assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.EQUAL)])) == len(
-            default_data_frame[default_data_frame[COLUMN_NAME_1] == 1]
-        )
-        assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.NOT_EQUAL)])) == len(
-            default_data_frame[default_data_frame[COLUMN_NAME_1] != 1]
-        )
-        assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.LESS_THAN)])) == len(
-            default_data_frame[default_data_frame[COLUMN_NAME_1] < 1]
-        )
-        assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.LESS_OR_EQUAL)])) == len(
-            default_data_frame[default_data_frame[COLUMN_NAME_1] <= 1]
-        )
-        assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.GREATER_THAN)])) == len(
-            default_data_frame[default_data_frame[COLUMN_NAME_1] > 1]
-        )
-        assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.GREATER_OR_EQUAL)])) == len(
-            default_data_frame[default_data_frame[COLUMN_NAME_1] >= 1]
-        )
-        assert len(df_dn.filter([(COLUMN_NAME_1, -1000, Operator.LESS_OR_EQUAL)])) == 0
-        assert len(df_dn.filter([(COLUMN_NAME_1, 1000, Operator.GREATER_OR_EQUAL)])) == 0
-        assert len(df_dn.filter([(COLUMN_NAME_1, 4, Operator.EQUAL), (COLUMN_NAME_1, 5, Operator.EQUAL)])) == len(
-            default_data_frame[(default_data_frame[COLUMN_NAME_1] == 4) & (default_data_frame[COLUMN_NAME_1] == 5)]
-        )
-        assert len(
-            df_dn.filter([(COLUMN_NAME_1, 4, Operator.EQUAL), (COLUMN_NAME_2, 5, Operator.EQUAL)], JoinOperator.OR)
-        ) == len(
-            default_data_frame[(default_data_frame[COLUMN_NAME_1] == 4) | (default_data_frame[COLUMN_NAME_2] == 5)]
-        )
-        assert len(
-            df_dn.filter(
-                [(COLUMN_NAME_1, 1, Operator.GREATER_THAN), (COLUMN_NAME_2, 3, Operator.GREATER_THAN)], JoinOperator.AND
-            )
-        ) == len(default_data_frame[(default_data_frame[COLUMN_NAME_1] > 1) & (default_data_frame[COLUMN_NAME_2] > 3)])
-        assert len(
-            df_dn.filter(
-                [(COLUMN_NAME_1, 2, Operator.GREATER_THAN), (COLUMN_NAME_1, 3, Operator.GREATER_THAN)], JoinOperator.OR
-            )
-        ) == len(default_data_frame[(default_data_frame[COLUMN_NAME_1] > 2) | (default_data_frame[COLUMN_NAME_1] > 3)])
-        assert len(
-            df_dn.filter(
-                [(COLUMN_NAME_1, 10, Operator.GREATER_THAN), (COLUMN_NAME_1, -10, Operator.LESS_THAN)], JoinOperator.AND
-            )
-        ) == len(
-            default_data_frame[(default_data_frame[COLUMN_NAME_1] > 10) | (default_data_frame[COLUMN_NAME_1] < -10)]
-        )
-        assert len(
-            df_dn.filter(
-                [(COLUMN_NAME_1, 10, Operator.GREATER_THAN), (COLUMN_NAME_1, -10, Operator.LESS_THAN)], JoinOperator.OR
-            )
-        ) == len(
-            default_data_frame[(default_data_frame[COLUMN_NAME_1] > 10) | (default_data_frame[COLUMN_NAME_1] < -10)]
-        )
-
-    def test_filter_list(self):
-        list_dn = FakeListDataNode("fake_list_dn")
-
-        KEY_NAME = "value"
-
-        assert len(list_dn.filter((KEY_NAME, 4, Operator.EQUAL))) == 1
-        assert len(list_dn.filter((KEY_NAME, 4, Operator.NOT_EQUAL))) == 9
-        assert len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL)])) == 1
-        assert len(list_dn.filter([(KEY_NAME, 4, Operator.NOT_EQUAL)])) == 9
-        assert len(list_dn.filter([(KEY_NAME, 4, Operator.LESS_THAN)])) == 4
-        assert len(list_dn.filter([(KEY_NAME, 4, Operator.LESS_OR_EQUAL)])) == 5
-        assert len(list_dn.filter([(KEY_NAME, 4, Operator.GREATER_THAN)])) == 5
-        assert len(list_dn.filter([(KEY_NAME, 4, Operator.GREATER_OR_EQUAL)])) == 6
-        assert len(list_dn.filter([(KEY_NAME, -1000, Operator.LESS_OR_EQUAL)])) == 0
-        assert len(list_dn.filter([(KEY_NAME, 1000, Operator.GREATER_OR_EQUAL)])) == 0
-
-        assert len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL), (KEY_NAME, 5, Operator.EQUAL)])) == 0
-        assert len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL), (KEY_NAME, 5, Operator.EQUAL)], JoinOperator.OR)) == 2
-        assert (
-            len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL), (KEY_NAME, 11, Operator.EQUAL)], JoinOperator.AND)) == 0
-        )
-        assert (
-            len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL), (KEY_NAME, 11, Operator.EQUAL)], JoinOperator.OR)) == 1
-        )
-
-        assert (
-            len(list_dn.filter([(KEY_NAME, -10, Operator.LESS_OR_EQUAL), (KEY_NAME, 11, Operator.GREATER_OR_EQUAL)]))
-            == 0
-        )
-        assert (
-            len(
-                list_dn.filter(
-                    [
-                        (KEY_NAME, 4, Operator.GREATER_OR_EQUAL),
-                        (KEY_NAME, 6, Operator.GREATER_OR_EQUAL),
-                    ],
-                    JoinOperator.AND,
-                )
-            )
-            == 4
-        )
-        assert (
-            len(
-                list_dn.filter(
-                    [
-                        (KEY_NAME, 4, Operator.GREATER_OR_EQUAL),
-                        (KEY_NAME, 6, Operator.GREATER_OR_EQUAL),
-                        (KEY_NAME, 11, Operator.EQUAL),
-                    ],
-                    JoinOperator.AND,
-                )
-            )
-            == 0
-        )
-        assert (
-            len(
-                list_dn.filter(
-                    [
-                        (KEY_NAME, 4, Operator.GREATER_OR_EQUAL),
-                        (KEY_NAME, 6, Operator.GREATER_OR_EQUAL),
-                        (KEY_NAME, 11, Operator.EQUAL),
-                    ],
-                    JoinOperator.OR,
-                )
-            )
-            == 6
-        )
-
-    def test_filter_numpy_exposed_type(self, default_data_frame):
-        default_array = default_data_frame.to_numpy()
-
-        df_dn = FakeNumpyarrayDataNode("fake_dataframe_dn", default_array)
-
-        assert len(df_dn.filter((0, 1, Operator.EQUAL))) == len(default_array[default_array[:, 0] == 1])
-        assert len(df_dn.filter((0, 1, Operator.NOT_EQUAL))) == len(default_array[default_array[:, 0] != 1])
-        assert len(df_dn.filter([(0, 1, Operator.EQUAL)])) == len(default_array[default_array[:, 0] == 1])
-        assert len(df_dn.filter([(0, 1, Operator.NOT_EQUAL)])) == len(default_array[default_array[:, 0] != 1])
-        assert len(df_dn.filter([(0, 1, Operator.LESS_THAN)])) == len(default_array[default_array[:, 0] < 1])
-        assert len(df_dn.filter([(0, 1, Operator.LESS_OR_EQUAL)])) == len(default_array[default_array[:, 0] <= 1])
-        assert len(df_dn.filter([(0, 1, Operator.GREATER_THAN)])) == len(default_array[default_array[:, 0] > 1])
-        assert len(df_dn.filter([(0, 1, Operator.GREATER_OR_EQUAL)])) == len(default_array[default_array[:, 0] >= 1])
-        assert len(df_dn.filter([(0, -1000, Operator.LESS_OR_EQUAL)])) == 0
-        assert len(df_dn.filter([(0, 1000, Operator.GREATER_OR_EQUAL)])) == 0
-        assert len(df_dn.filter([(0, 4, Operator.EQUAL), (0, 5, Operator.EQUAL)])) == len(
-            default_array[(default_array[:, 0] == 4) & (default_array[:, 0] == 5)]
-        )
-        assert len(df_dn.filter([(0, 4, Operator.EQUAL), (1, 5, Operator.EQUAL)], JoinOperator.OR)) == len(
-            default_array[(default_array[:, 0] == 4) | (default_array[:, 1] == 5)]
-        )
-        assert len(
-            df_dn.filter([(0, 1, Operator.GREATER_THAN), (1, 3, Operator.GREATER_THAN)], JoinOperator.AND)
-        ) == len(default_array[(default_array[:, 0] > 1) & (default_array[:, 1] > 3)])
-        assert len(
-            df_dn.filter([(0, 2, Operator.GREATER_THAN), (0, 3, Operator.GREATER_THAN)], JoinOperator.OR)
-        ) == len(default_array[(default_array[:, 0] > 2) | (default_array[:, 0] > 3)])
-        assert len(
-            df_dn.filter([(0, 10, Operator.GREATER_THAN), (0, -10, Operator.LESS_THAN)], JoinOperator.AND)
-        ) == len(default_array[(default_array[:, 0] > 10) | (default_array[:, 0] < -10)])
-        assert len(
-            df_dn.filter([(0, 10, Operator.GREATER_THAN), (0, -10, Operator.LESS_THAN)], JoinOperator.OR)
-        ) == len(default_array[(default_array[:, 0] > 10) | (default_array[:, 0] < -10)])
 
     def test_data_node_update_after_writing(self):
         dn = FakeDataNode("foo")
