+++ tests/core/data/test_filter_data_node.py
@@ -13,392 +13,271 @@ from typing import Dict, List
 
 import numpy as np
 import pandas as pd
+import pytest
 
-from src.taipy.core.data._filter import _FilterDataNode
-from src.taipy.core.data.data_node import DataNode
+from src.taipy.core.data.operator import JoinOperator, Operator
+
+from .utils import (
+    CustomClass,
+    FakeCustomDataNode,
+    FakeDataframeDataNode,
+    FakeDataNode,
+    FakeListDataNode,
+    FakeMultiSheetExcelCustomDataNode,
+    FakeMultiSheetExcelDataFrameDataNode,
+    FakeNumpyarrayDataNode,
+)
 
 
-class FakeDataframeDataNode(DataNode):
+def test_filter_pandas_exposed_type(default_data_frame):
+    dn = FakeDataNode("fake_dn")
+    dn.write("Any data")
+
+    with pytest.raises(NotImplementedError):
+        dn.filter((("any", 0, Operator.EQUAL)), JoinOperator.OR)
+    with pytest.raises(NotImplementedError):
+        dn.filter((("any", 0, Operator.NOT_EQUAL)), JoinOperator.OR)
+    with pytest.raises(NotImplementedError):
+        dn.filter((("any", 0, Operator.LESS_THAN)), JoinOperator.AND)
+    with pytest.raises(NotImplementedError):
+        dn.filter((("any", 0, Operator.LESS_OR_EQUAL)), JoinOperator.AND)
+    with pytest.raises(NotImplementedError):
+        dn.filter((("any", 0, Operator.GREATER_THAN)))
+    with pytest.raises(NotImplementedError):
+        dn.filter(("any", 0, Operator.GREATER_OR_EQUAL))
+
+    df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
+
     COLUMN_NAME_1 = "a"
     COLUMN_NAME_2 = "b"
-
-    def __init__(self, config_id, default_data_frame, **kwargs):
-        super().__init__(config_id, **kwargs)
-        self.data = default_data_frame
-
-    def _read(self):
-        return self.data
-
-
-class CustomClass:
-    def __init__(self, a, b):
-        self.a = a
-        self.b = b
-
-
-class FakeCustomDataNode(DataNode):
-    def __init__(self, config_id, **kwargs):
-        super().__init__(config_id, **kwargs)
-        self.data = [CustomClass(i, i * 2) for i in range(10)]
-
-    def _read(self):
-        return self.data
-
-
-class FakeMultiSheetExcelDataFrameDataNode(DataNode):
-    def __init__(self, config_id, default_data_frame, **kwargs):
-        super().__init__(config_id, **kwargs)
-        self.data = {
-            "Sheet1": default_data_frame,
-            "Sheet2": default_data_frame,
-        }
-
-    def _read(self):
-        return self.data
-
-
-class FakeMultiSheetExcelCustomDataNode(DataNode):
-    def __init__(self, config_id, **kwargs):
-        super().__init__(config_id, **kwargs)
-        self.data = {
-            "Sheet1": [CustomClass(i, i * 2) for i in range(10)],
-            "Sheet2": [CustomClass(i, i * 2) for i in range(10)],
-        }
-
-    def _read(self):
-        return self.data
-
-
-class TestFilterDataNode:
-    def test_get_item(self, default_data_frame):
-        # get item for DataFrame data_type
-        default_data_frame[1] = [100, 100]
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
-
-        filtered_df_dn = df_dn["a"]
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert len(filtered_df_dn.data) == len(default_data_frame["a"])
-        assert filtered_df_dn.data.to_dict() == default_data_frame["a"].to_dict()
-
-        filtered_df_dn = df_dn[1]
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert len(filtered_df_dn.data) == len(default_data_frame[1])
-        assert filtered_df_dn.data.to_dict() == default_data_frame[1].to_dict()
-
-        filtered_df_dn = df_dn[0:2]
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert filtered_df_dn.data.shape == default_data_frame[0:2].shape
-        assert len(filtered_df_dn.data) == 2
-
-        bool_df = default_data_frame.copy(deep=True) > 4
-        filtered_df_dn = df_dn[bool_df]
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-
-        bool_1d_index = [True, False]
-        filtered_df_dn = df_dn[bool_1d_index]
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert filtered_df_dn.data.to_dict() == default_data_frame[bool_1d_index].to_dict()
-        assert len(filtered_df_dn.data) == 1
-
-        filtered_df_dn = df_dn[["a", "b"]]
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert filtered_df_dn.data.shape == default_data_frame[["a", "b"]].shape
-        assert filtered_df_dn.data.to_dict() == default_data_frame[["a", "b"]].to_dict()
-
-        # get item for custom data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
-
-        filtered_custom_dn = custom_dn["a"]
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert len(filtered_custom_dn.data) == 10
-        assert filtered_custom_dn.data == [i for i in range(10)]
-
-        filtered_custom_dn = custom_dn[0:5]
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, CustomClass) for x in filtered_custom_dn.data])
-        assert len(filtered_custom_dn.data) == 5
-
-        bool_df = pd.DataFrame({"a": [i for i in range(10)], "b": [i * 2 for i in range(10)]}) > 4
-        filtered_custom_dn = custom_dn[["a", "b"]][bool_df]
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, Dict) for x in filtered_custom_dn.data])
-        for i, row in bool_df.iterrows():
-            for col in row.index:
-                if row[col]:
-                    assert filtered_custom_dn.data[i][col] == custom_dn[["a", "b"]].data[i][col]
-                else:
-                    assert filtered_custom_dn.data[i][col] is None
-
-        filtered_custom_dn = custom_dn["a"][bool_df]
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert filtered_custom_dn.data is None
-
-        filtered_custom_dn = custom_dn[0:10][bool_df]
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert filtered_custom_dn.data is None
-
-        bool_1d_index = [True if i < 5 else False for i in range(10)]
-        filtered_custom_dn = custom_dn[bool_1d_index]
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert len(filtered_custom_dn.data) == 5
-        assert filtered_custom_dn.data == custom_dn._read()[:5]
-
-        filtered_custom_dn = custom_dn[["a", "b"]]
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, Dict) for x in filtered_custom_dn.data])
-        assert len(filtered_custom_dn.data) == 10
-        assert filtered_custom_dn.data == [{"a": i, "b": i * 2} for i in range(10)]
-
-        # get item for Multi-sheet Excel data_type
-        multi_sheet_excel_df_dn = FakeMultiSheetExcelDataFrameDataNode(
-            "fake_multi_sheet_excel_df_dn", default_data_frame
+    assert len(df_dn.filter((COLUMN_NAME_1, 1, Operator.EQUAL))) == len(
+        default_data_frame[default_data_frame[COLUMN_NAME_1] == 1]
+    )
+    assert len(df_dn.filter((COLUMN_NAME_1, 1, Operator.NOT_EQUAL))) == len(
+        default_data_frame[default_data_frame[COLUMN_NAME_1] != 1]
+    )
+    assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.EQUAL)])) == len(
+        default_data_frame[default_data_frame[COLUMN_NAME_1] == 1]
+    )
+    assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.NOT_EQUAL)])) == len(
+        default_data_frame[default_data_frame[COLUMN_NAME_1] != 1]
+    )
+    assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.LESS_THAN)])) == len(
+        default_data_frame[default_data_frame[COLUMN_NAME_1] < 1]
+    )
+    assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.LESS_OR_EQUAL)])) == len(
+        default_data_frame[default_data_frame[COLUMN_NAME_1] <= 1]
+    )
+    assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.GREATER_THAN)])) == len(
+        default_data_frame[default_data_frame[COLUMN_NAME_1] > 1]
+    )
+    assert len(df_dn.filter([(COLUMN_NAME_1, 1, Operator.GREATER_OR_EQUAL)])) == len(
+        default_data_frame[default_data_frame[COLUMN_NAME_1] >= 1]
+    )
+    assert len(df_dn.filter([(COLUMN_NAME_1, -1000, Operator.LESS_OR_EQUAL)])) == 0
+    assert len(df_dn.filter([(COLUMN_NAME_1, 1000, Operator.GREATER_OR_EQUAL)])) == 0
+    assert len(df_dn.filter([(COLUMN_NAME_1, 4, Operator.EQUAL), (COLUMN_NAME_1, 5, Operator.EQUAL)])) == len(
+        default_data_frame[(default_data_frame[COLUMN_NAME_1] == 4) & (default_data_frame[COLUMN_NAME_1] == 5)]
+    )
+    assert len(
+        df_dn.filter([(COLUMN_NAME_1, 4, Operator.EQUAL), (COLUMN_NAME_2, 5, Operator.EQUAL)], JoinOperator.OR)
+    ) == len(default_data_frame[(default_data_frame[COLUMN_NAME_1] == 4) | (default_data_frame[COLUMN_NAME_2] == 5)])
+    assert len(
+        df_dn.filter(
+            [(COLUMN_NAME_1, 1, Operator.GREATER_THAN), (COLUMN_NAME_2, 3, Operator.GREATER_THAN)], JoinOperator.AND
         )
-        filtered_multi_sheet_excel_df_dn = multi_sheet_excel_df_dn["Sheet1"]
-        assert isinstance(filtered_multi_sheet_excel_df_dn, _FilterDataNode)
-        assert isinstance(filtered_multi_sheet_excel_df_dn.data, pd.DataFrame)
-        assert len(filtered_multi_sheet_excel_df_dn.data) == len(default_data_frame)
-        assert np.array_equal(filtered_multi_sheet_excel_df_dn.data.to_numpy(), default_data_frame.to_numpy())
-
-        multi_sheet_excel_custom_dn = FakeMultiSheetExcelCustomDataNode("fake_multi_sheet_excel_df_dn")
-        filtered_multi_sheet_excel_custom_dn = multi_sheet_excel_custom_dn["Sheet1"]
-        assert isinstance(filtered_multi_sheet_excel_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_multi_sheet_excel_custom_dn.data, List)
-        assert len(filtered_multi_sheet_excel_custom_dn.data) == 10
-        expected_value = [CustomClass(i, i * 2) for i in range(10)]
-        assert all(
-            [
-                expected.a == filtered.a and expected.b == filtered.b
-                for expected, filtered in zip(expected_value, filtered_multi_sheet_excel_custom_dn.data)
-            ]
+    ) == len(default_data_frame[(default_data_frame[COLUMN_NAME_1] > 1) & (default_data_frame[COLUMN_NAME_2] > 3)])
+    assert len(
+        df_dn.filter(
+            [(COLUMN_NAME_1, 2, Operator.GREATER_THAN), (COLUMN_NAME_1, 3, Operator.GREATER_THAN)], JoinOperator.OR
         )
+    ) == len(default_data_frame[(default_data_frame[COLUMN_NAME_1] > 2) | (default_data_frame[COLUMN_NAME_1] > 3)])
+    assert len(
+        df_dn.filter(
+            [(COLUMN_NAME_1, 10, Operator.GREATER_THAN), (COLUMN_NAME_1, -10, Operator.LESS_THAN)], JoinOperator.AND
+        )
+    ) == len(default_data_frame[(default_data_frame[COLUMN_NAME_1] > 10) | (default_data_frame[COLUMN_NAME_1] < -10)])
+    assert len(
+        df_dn.filter(
+            [(COLUMN_NAME_1, 10, Operator.GREATER_THAN), (COLUMN_NAME_1, -10, Operator.LESS_THAN)], JoinOperator.OR
+        )
+    ) == len(default_data_frame[(default_data_frame[COLUMN_NAME_1] > 10) | (default_data_frame[COLUMN_NAME_1] < -10)])
 
-    def test_equal(self, default_data_frame):
-        # equal to for pandas dataframe data_type
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
 
-        filtered_df_dn = df_dn["a"] == 1
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert filtered_df_dn.data.dtype == bool
-        assert all(filtered_df_dn.data == (default_data_frame["a"] == 1))
+def test_filter_list():
+    list_dn = FakeListDataNode("fake_list_dn")
 
-        filtered_df_dn = df_dn[["a", "b"]] == 1
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert all(filtered_df_dn.data.dtypes == bool)
-        assert all(filtered_df_dn.data == (default_data_frame[["a", "b"]] == 1))
+    KEY_NAME = "value"
 
-        # equal to for custom list data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
+    assert len(list_dn.filter((KEY_NAME, 4, Operator.EQUAL))) == 1
+    assert len(list_dn.filter((KEY_NAME, 4, Operator.NOT_EQUAL))) == 9
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL)])) == 1
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.NOT_EQUAL)])) == 9
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.LESS_THAN)])) == 4
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.LESS_OR_EQUAL)])) == 5
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.GREATER_THAN)])) == 5
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.GREATER_OR_EQUAL)])) == 6
+    assert len(list_dn.filter([(KEY_NAME, -1000, Operator.LESS_OR_EQUAL)])) == 0
+    assert len(list_dn.filter([(KEY_NAME, 1000, Operator.GREATER_OR_EQUAL)])) == 0
 
-        filtered_custom_dn = custom_dn["a"] == 0
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [True] + [False for _ in range(9)]
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL), (KEY_NAME, 5, Operator.EQUAL)])) == 0
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL), (KEY_NAME, 5, Operator.EQUAL)], JoinOperator.OR)) == 2
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL), (KEY_NAME, 11, Operator.EQUAL)], JoinOperator.AND)) == 0
+    assert len(list_dn.filter([(KEY_NAME, 4, Operator.EQUAL), (KEY_NAME, 11, Operator.EQUAL)], JoinOperator.OR)) == 1
 
-        filtered_custom_dn = custom_dn[["a", "b"]] == 0
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [False for _ in range(10)]
+    assert (
+        len(list_dn.filter([(KEY_NAME, -10, Operator.LESS_OR_EQUAL), (KEY_NAME, 11, Operator.GREATER_OR_EQUAL)])) == 0
+    )
+    assert (
+        len(
+            list_dn.filter(
+                [
+                    (KEY_NAME, 4, Operator.GREATER_OR_EQUAL),
+                    (KEY_NAME, 6, Operator.GREATER_OR_EQUAL),
+                ],
+                JoinOperator.AND,
+            )
+        )
+        == 4
+    )
+    assert (
+        len(
+            list_dn.filter(
+                [
+                    (KEY_NAME, 4, Operator.GREATER_OR_EQUAL),
+                    (KEY_NAME, 6, Operator.GREATER_OR_EQUAL),
+                    (KEY_NAME, 11, Operator.EQUAL),
+                ],
+                JoinOperator.AND,
+            )
+        )
+        == 0
+    )
+    assert (
+        len(
+            list_dn.filter(
+                [
+                    (KEY_NAME, 4, Operator.GREATER_OR_EQUAL),
+                    (KEY_NAME, 6, Operator.GREATER_OR_EQUAL),
+                    (KEY_NAME, 11, Operator.EQUAL),
+                ],
+                JoinOperator.OR,
+            )
+        )
+        == 6
+    )
 
-    def test_not_equal(self, default_data_frame):
-        # not equal to for pandas dataframe data_type
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
 
-        filtered_df_dn = df_dn["a"] != 1
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert filtered_df_dn.data.dtype == bool
-        assert all(filtered_df_dn.data == (default_data_frame["a"] != 1))
+def test_filter_numpy_exposed_type(default_data_frame):
+    default_array = default_data_frame.to_numpy()
 
-        filtered_df_dn = df_dn[["a", "b"]] != 1
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert all(filtered_df_dn.data.dtypes == bool)
-        assert all(filtered_df_dn.data == (default_data_frame[["a", "b"]] != 1))
+    df_dn = FakeNumpyarrayDataNode("fake_dataframe_dn", default_array)
 
-        # not equal to for custom list data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
+    assert len(df_dn.filter((0, 1, Operator.EQUAL))) == len(default_array[default_array[:, 0] == 1])
+    assert len(df_dn.filter((0, 1, Operator.NOT_EQUAL))) == len(default_array[default_array[:, 0] != 1])
+    assert len(df_dn.filter([(0, 1, Operator.EQUAL)])) == len(default_array[default_array[:, 0] == 1])
+    assert len(df_dn.filter([(0, 1, Operator.NOT_EQUAL)])) == len(default_array[default_array[:, 0] != 1])
+    assert len(df_dn.filter([(0, 1, Operator.LESS_THAN)])) == len(default_array[default_array[:, 0] < 1])
+    assert len(df_dn.filter([(0, 1, Operator.LESS_OR_EQUAL)])) == len(default_array[default_array[:, 0] <= 1])
+    assert len(df_dn.filter([(0, 1, Operator.GREATER_THAN)])) == len(default_array[default_array[:, 0] > 1])
+    assert len(df_dn.filter([(0, 1, Operator.GREATER_OR_EQUAL)])) == len(default_array[default_array[:, 0] >= 1])
+    assert len(df_dn.filter([(0, -1000, Operator.LESS_OR_EQUAL)])) == 0
+    assert len(df_dn.filter([(0, 1000, Operator.GREATER_OR_EQUAL)])) == 0
+    assert len(df_dn.filter([(0, 4, Operator.EQUAL), (0, 5, Operator.EQUAL)])) == len(
+        default_array[(default_array[:, 0] == 4) & (default_array[:, 0] == 5)]
+    )
+    assert len(df_dn.filter([(0, 4, Operator.EQUAL), (1, 5, Operator.EQUAL)], JoinOperator.OR)) == len(
+        default_array[(default_array[:, 0] == 4) | (default_array[:, 1] == 5)]
+    )
+    assert len(df_dn.filter([(0, 1, Operator.GREATER_THAN), (1, 3, Operator.GREATER_THAN)], JoinOperator.AND)) == len(
+        default_array[(default_array[:, 0] > 1) & (default_array[:, 1] > 3)]
+    )
+    assert len(df_dn.filter([(0, 2, Operator.GREATER_THAN), (0, 3, Operator.GREATER_THAN)], JoinOperator.OR)) == len(
+        default_array[(default_array[:, 0] > 2) | (default_array[:, 0] > 3)]
+    )
+    assert len(df_dn.filter([(0, 10, Operator.GREATER_THAN), (0, -10, Operator.LESS_THAN)], JoinOperator.AND)) == len(
+        default_array[(default_array[:, 0] > 10) | (default_array[:, 0] < -10)]
+    )
+    assert len(df_dn.filter([(0, 10, Operator.GREATER_THAN), (0, -10, Operator.LESS_THAN)], JoinOperator.OR)) == len(
+        default_array[(default_array[:, 0] > 10) | (default_array[:, 0] < -10)]
+    )
 
-        filtered_custom_dn = custom_dn["a"] != 0
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [False] + [True for _ in range(9)]
 
-        filtered_custom_dn = custom_dn[["a", "b"]] != 0
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [True for _ in range(10)]
+def test_filter_by_get_item(default_data_frame):
+    # get item for DataFrame data_type
+    default_data_frame[1] = [100, 100]
+    df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
 
-    def test_larger_than(self, default_data_frame):
-        # larger than for pandas dataframe data_type
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
+    filtered_df_dn = df_dn["a"]
+    assert isinstance(filtered_df_dn, pd.Series)
+    assert len(filtered_df_dn) == len(default_data_frame["a"])
+    assert filtered_df_dn.to_dict() == default_data_frame["a"].to_dict()
 
-        filtered_df_dn = df_dn["a"] > 2
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert filtered_df_dn.data.dtype == bool
-        assert all(filtered_df_dn.data == (default_data_frame["a"] > 2))
+    filtered_df_dn = df_dn[1]
+    assert isinstance(filtered_df_dn, pd.Series)
+    assert len(filtered_df_dn) == len(default_data_frame[1])
+    assert filtered_df_dn.to_dict() == default_data_frame[1].to_dict()
 
-        filtered_df_dn = df_dn[["a", "b"]] > 2
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert all(filtered_df_dn.data.dtypes == bool)
-        assert all(filtered_df_dn.data == (default_data_frame[["a", "b"]] > 2))
+    filtered_df_dn = df_dn[0:2]
+    assert isinstance(filtered_df_dn, pd.DataFrame)
+    assert filtered_df_dn.shape == default_data_frame[0:2].shape
+    assert len(filtered_df_dn) == 2
 
-        # larger than for custom data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
+    bool_df = default_data_frame.copy(deep=True) > 4
+    filtered_df_dn = df_dn[bool_df]
+    assert isinstance(filtered_df_dn, pd.DataFrame)
 
-        filtered_custom_dn = custom_dn["a"] > 5
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [False for _ in range(6)] + [True for _ in range(4)]
+    bool_1d_index = [True, False]
+    filtered_df_dn = df_dn[bool_1d_index]
+    assert isinstance(filtered_df_dn, pd.DataFrame)
+    assert filtered_df_dn.to_dict() == default_data_frame[bool_1d_index].to_dict()
+    assert len(filtered_df_dn) == 1
 
-    def test_larger_equal_to(self, default_data_frame):
-        # larger than or equal to for pandas dataframe data_type
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
+    filtered_df_dn = df_dn[["a", "b"]]
+    assert isinstance(filtered_df_dn, pd.DataFrame)
+    assert filtered_df_dn.shape == default_data_frame[["a", "b"]].shape
+    assert filtered_df_dn.to_dict() == default_data_frame[["a", "b"]].to_dict()
 
-        filtered_df_dn = df_dn["a"] >= 4
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert filtered_df_dn.data.dtype == bool
-        assert all(filtered_df_dn.data == (default_data_frame["a"] >= 4))
+    # get item for custom data_type
+    custom_dn = FakeCustomDataNode("fake_custom_dn")
 
-        filtered_df_dn = df_dn[["a", "b"]] >= 4
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert all(filtered_df_dn.data.dtypes == bool)
-        assert all(filtered_df_dn.data == (default_data_frame[["a", "b"]] >= 4))
+    filtered_custom_dn = custom_dn["a"]
+    assert isinstance(filtered_custom_dn, List)
+    assert len(filtered_custom_dn) == 10
+    assert filtered_custom_dn == [i for i in range(10)]
 
-        # larger than or equal to for custom data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
+    filtered_custom_dn = custom_dn[0:5]
+    assert isinstance(filtered_custom_dn, List)
+    assert all([isinstance(x, CustomClass) for x in filtered_custom_dn])
+    assert len(filtered_custom_dn) == 5
 
-        filtered_custom_dn = custom_dn["a"] >= 5
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [False for _ in range(5)] + [True for _ in range(5)]
+    bool_1d_index = [True if i < 5 else False for i in range(10)]
+    filtered_custom_dn = custom_dn[bool_1d_index]
+    assert isinstance(filtered_custom_dn, List)
+    assert len(filtered_custom_dn) == 5
+    assert filtered_custom_dn == custom_dn._read()[:5]
 
-    def test_lesser_than(self, default_data_frame):
-        # lesser than for pandas dataframe data_type
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
+    filtered_custom_dn = custom_dn[["a", "b"]]
+    assert isinstance(filtered_custom_dn, List)
+    assert all([isinstance(x, Dict) for x in filtered_custom_dn])
+    assert len(filtered_custom_dn) == 10
+    assert filtered_custom_dn == [{"a": i, "b": i * 2} for i in range(10)]
 
-        filtered_df_dn = df_dn["a"] < 5
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert filtered_df_dn.data.dtype == bool
-        assert all(filtered_df_dn.data == (default_data_frame["a"] < 5))
+    # get item for Multi-sheet Excel data_type
+    multi_sheet_excel_df_dn = FakeMultiSheetExcelDataFrameDataNode("fake_multi_sheet_excel_df_dn", default_data_frame)
+    filtered_multi_sheet_excel_df_dn = multi_sheet_excel_df_dn["Sheet1"]
+    assert isinstance(filtered_multi_sheet_excel_df_dn, pd.DataFrame)
+    assert len(filtered_multi_sheet_excel_df_dn) == len(default_data_frame)
+    assert np.array_equal(filtered_multi_sheet_excel_df_dn.to_numpy(), default_data_frame.to_numpy())
 
-        filtered_df_dn = df_dn[["a", "b"]] < 5
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert all(filtered_df_dn.data.dtypes == bool)
-        assert all(filtered_df_dn.data == (default_data_frame[["a", "b"]] < 5))
-
-        # lesser than for custom data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
-
-        filtered_custom_dn = custom_dn["a"] < 5
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [True for _ in range(5)] + [False for _ in range(5)]
-
-    def test_lesser_equal_to(self, default_data_frame):
-        # lesser than or equal to for pandas dataframe data_type
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
-
-        filtered_df_dn = df_dn["a"] <= 5
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert filtered_df_dn.data.dtype == bool
-        assert all(filtered_df_dn.data == (default_data_frame["a"] <= 5))
-
-        filtered_df_dn = df_dn[["a", "b"]] <= 5
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert all(filtered_df_dn.data.dtypes == bool)
-        assert all(filtered_df_dn.data == (default_data_frame[["a", "b"]] <= 5))
-
-        # lesser than or equal to for custom data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
-
-        filtered_custom_dn = custom_dn["a"] <= 5
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [True for _ in range(6)] + [False for _ in range(4)]
-
-    def test_and(self, default_data_frame):
-        # and comparator for pandas dataframe data_type
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
-
-        filtered_df_dn = (df_dn["a"] >= 2) & (df_dn["a"] <= 5)
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert filtered_df_dn.data.dtype == bool
-        assert all(filtered_df_dn.data == (default_data_frame["a"] >= 2) & (default_data_frame["a"] <= 5))
-
-        filtered_df_dn = (df_dn[["a", "b"]] >= 2) & (df_dn[["a", "b"]] <= 5)
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert all(filtered_df_dn.data.dtypes == bool)
-        assert all(filtered_df_dn.data == (default_data_frame[["a", "b"]] >= 2) & (default_data_frame[["a", "b"]] <= 5))
-
-        # and comparator for custom data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
-
-        filtered_custom_dn = (custom_dn["a"] >= 2) & (custom_dn["a"] <= 5)
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [False for _ in range(2)] + [True for _ in range(4)] + [
-            False for _ in range(4)
+    multi_sheet_excel_custom_dn = FakeMultiSheetExcelCustomDataNode("fake_multi_sheet_excel_df_dn")
+    filtered_multi_sheet_excel_custom_dn = multi_sheet_excel_custom_dn["Sheet1"]
+    assert isinstance(filtered_multi_sheet_excel_custom_dn, List)
+    assert len(filtered_multi_sheet_excel_custom_dn) == 10
+    expected_value = [CustomClass(i, i * 2) for i in range(10)]
+    assert all(
+        [
+            expected.a == filtered.a and expected.b == filtered.b
+            for expected, filtered in zip(expected_value, filtered_multi_sheet_excel_custom_dn)
         ]
-
-    def test_or(self, default_data_frame):
-        # or comparator for pandas dataframe data_type
-        df_dn = FakeDataframeDataNode("fake_dataframe_dn", default_data_frame)
-
-        filtered_df_dn = (df_dn["a"] < 2) | (df_dn["a"] > 5)
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.Series)
-        assert filtered_df_dn.data.dtype == bool
-        assert all(filtered_df_dn.data == (default_data_frame["a"] < 2) | (default_data_frame["a"] > 5))
-
-        filtered_df_dn = (df_dn[["a", "b"]] < 2) | (df_dn[["a", "b"]] > 5)
-        assert isinstance(filtered_df_dn, _FilterDataNode)
-        assert isinstance(filtered_df_dn.data, pd.DataFrame)
-        assert all(filtered_df_dn.data.dtypes == bool)
-        assert all(filtered_df_dn.data == (default_data_frame[["a", "b"]] < 2) | (default_data_frame[["a", "b"]] > 5))
-
-        # or comparator for custom data_type
-        custom_dn = FakeCustomDataNode("fake_custom_dn")
-
-        filtered_custom_dn = (custom_dn["a"] < 2) | (custom_dn["a"] > 5)
-        assert isinstance(filtered_custom_dn, _FilterDataNode)
-        assert isinstance(filtered_custom_dn.data, List)
-        assert all([isinstance(x, bool) for x in filtered_custom_dn.data])
-        assert filtered_custom_dn.data == [True for _ in range(2)] + [False for _ in range(4)] + [
-            True for _ in range(4)
-        ]
-
-    def test_to_string(self):
-        filter_dn = _FilterDataNode("dn_id", [])
-        assert isinstance(str(filter_dn), str)
+    )
